upgrade template and install
