--- packages/soul-engine-cli/template/soul/lib/cognitiveSteps.ts
@@ -1,124 +1,0 @@
-import { createCognitiveStep, ChatMessageRoleEnum, WorkingMemory, indentNicely, stripEntityAndVerb, stripEntityAndVerbFromStream, z  } from "@opensouls/engine";
-
-export const externalDialog = createCognitiveStep((instructions: string | { instructions: string; verb: string }) => {
-  let instructionString: string, verb: string;
-  if (typeof instructions === "string") {
-    instructionString = instructions;
-    verb = "said";
-  } else {
-    instructionString = instructions.instructions;
-    verb = instructions.verb;
-  }
-  return {
-    command: ({ soulName: name }: WorkingMemory) => {
-      return {
-        role: ChatMessageRoleEnum.System,
-        name: name,
-        content: indentNicely`
-          Model the mind of ${name}.
-    
-          ## Instructions
-          * DO NOT include actions (for example, do NOT add non-verbal items like *John Smiles* or *John Nods*, etc).
-          * DO NOT include internal thoughts (for example, do NOT respond with John thought: "...").
-          * If necessary, use all CAPS to emphasize certain words.
-    
-          ${instructionString}
-    
-          Please reply with the next utterance from ${name}. Use the format '${name} ${verb}: "..."'
-        `
-      };
-    },
-    streamProcessor: stripEntityAndVerbFromStream,
-    postProcess: async (memory: WorkingMemory, response: string) => {
-      const stripped = stripEntityAndVerb(memory.soulName, verb, response);
-      const newMemory = {
-        role: ChatMessageRoleEnum.Assistant,
-        content: `${memory.soulName} ${verb}: "${stripped}"`
-      };
-      return [newMemory, stripped];
-    }
-  }
-})
-
-export const internalMonologue = createCognitiveStep((instructions: string | { instructions: string; verb: string }) => {
-  let instructionString: string, verb: string;
-  if (typeof instructions === "string") {
-    instructionString = instructions;
-    verb = "thought";
-  } else {
-    instructionString = instructions.instructions;
-    verb = instructions.verb;
-  }
-
-  return {
-    command: ({ soulName: name }: WorkingMemory) => {
-      return {
-        role: ChatMessageRoleEnum.System,
-        name: name,
-        content: indentNicely`
-          Model the mind of ${name}.
-
-          ## Description
-          ${instructionString}
-
-          ## Rules
-          * Internal monologue thoughts should match the speaking style of ${name}.
-          * Only respond with the format '${name} ${verb}: "..."', no additional commentary or text.
-          * Follow the Description when creating the internal thought!
-
-          Please reply with the next internal monologue thought of ${name}. Use the format '${name} ${verb}: "..."'
-        `
-      };
-    },
-    streamProcessor: stripEntityAndVerbFromStream,
-    postProcess: async (memory: WorkingMemory, response: string) => {
-      const stripped = stripEntityAndVerb(memory.soulName, verb, response);
-      const newMemory = {
-        role: ChatMessageRoleEnum.Assistant,
-        content: `${memory.soulName} ${verb}: "${stripped}"`
-      };
-      return [newMemory, stripped];
-    }
-  }
-})
-
-export const mentalQuery = createCognitiveStep((statement: string) => {
-  const params = z.object({
-    isStatementTrue: z.boolean().describe(`Is the statement true or false?`),
-  });
-
-  return {
-    command: ({ soulName: name }: WorkingMemory) => {
-      return {
-        role: ChatMessageRoleEnum.System,
-        name: name,
-        content: indentNicely`
-          ${name} reasons about the veracity of the following statement.
-          > ${statement}
-
-          Please reply with if ${name} believes the statement is true or false.
-        `,
-      };
-    },
-    schema: params,
-    postProcess: async (memory: WorkingMemory, response: z.output<typeof params>) => {
-      const newMemory = {
-        role: ChatMessageRoleEnum.Assistant,
-        content: `${memory.soulName} evaluated: \`${statement}\` and decided that the statement is ${response.isStatementTrue ? 'true' : 'false'}`
-      };
-      return [newMemory, response.isStatementTrue];
-    }
-  };
-});
-
-export const instruction = createCognitiveStep((instructions: string) => {
-  return {
-    command: ({ soulName }: WorkingMemory) => {
-      return {
-        role: ChatMessageRoleEnum.System,
-        name: soulName,
-        content: instructions,
-      };
-    }
-  };
-});
